Fix cast exception

+++ Events/FSSDiscoveryScanEvent.cs
@@ -29,7 +29,7 @@ namespace EddiEvents
 
         public DiscoveryScanEvent(DateTime timestamp, decimal progress, int totalbodies, int nonbodies) : base(timestamp, NAME)
         {
-            this.progress = Math.Round(progress * 100); // multiplied by 100 to convert to percentage
+            this.progress = (int)Math.Round(progress * 100); // multiplied by 100 to convert to percentage
             this.totalbodies = totalbodies;
             this.nonbodies = nonbodies;
         }
